+++ src/Engine/Graphics/VertexData.h
@@ -61,7 +61,7 @@ public:
     void resetMeshData();
     void updateMesh(PMesh mesh, Component::Transform& transform);
     void createDescriptors();
-    void loadMesh(MeshId id, Array<uint16> indices, Array<Meshlet> meshlets);
+    void loadMesh(MeshId id, Array<uint32> indices, Array<Meshlet> meshlets);
     MeshId allocateVertexData(uint64 numVertices);
     uint64 getMeshOffset(MeshId id);
     uint64 getMeshVertexCount(MeshId id);
@@ -100,8 +100,8 @@ protected:
     Map<MeshId, uint64> meshVertexCounts;
     Array<MeshletDescription> meshlets;
     Array<uint8> primitiveIndices;
-    Array<uint16> vertexIndices;
-    Array<uint16> indices;
+    Array<uint32> vertexIndices;
+    Array<uint32> indices;
     Gfx::PGraphics graphics;
     Gfx::ODescriptorLayout instanceDataLayout;
     // for mesh shading
